The launch scripts now work!
The problem was that TCP and UDP checksum offloading was on in
the ous container.  The kernel expected the physical interface
to compute the checksum, but it's a veth, so there isn't one.
Therefore the checksum was wrong when it left the ous container,
and the packet treated as invalid and not sent though the nat table
later on.